BUG: Made num_params a float instead of int

+++ engine.py
@@ -126,7 +126,7 @@ if __name__ == '__main__':
     parser.add_argument('--p', type=int, default=1, help='Default p value for model')
     parser.add_argument('--k', type=int, default=2, help='Default k value for model')
     parser.add_argument('--g', type=int, default=1, help='Default g value for model')
-    parser.add_argument('--num_params', type=int, default=0, help='Adjust number of model params')
+    parser.add_argument('--num_params', type=float, default=0, help='Adjust number of model params')
     parser.add_argument('--resnet_ver', type=int, default=34, help='Which version of ResNet to use')
     parser.add_argument('--resnet_width', type=float, default=2, help='How wide to make our ResNet layers')
     parser.add_argument('--model', type=str, default='mlp', help='cnn, mlp, resnet')  # cnn
